Add support for per-package install_options configuration
This allows individual packages to be configured with custom
install_options. This works for both packages installed directly and
packages installed as dependencies of other packages.

For example:

    build.packages.config['foo'] = {
        'install_options': ['--no-compile']
    }

If the package "foo" were scheduled for installation (either directly or
as a dependency of another package), this configuration would cause the
package "foo" to be installed with the --no-compile installation option.