Improve package post-install version calculation performance
After installing packages, uranium stores the installed version of each
package to provide as constraints to successive package installation
commands.

The method pip uses to retrieve the version of an installed package is
inefficient, which often results in noticeable delay (on the order of
seconds for each invocation of build.packages.install).

This change reduces the frequency of retrieving each package's installed
version during the package post-installation phase with the following:

- Packages that were not installed as a result of the preceding install
  command are skipped (as the installed version of any prior installed
  packages should already be stored in the versions cache)

- If a constraint specifier was computed during package dependency
  resolution, use that constraint specifier instead of
  retrieving the installed version

Typically, the vast majority of installed packages are dependencies with
constraints attached, requiring installed version retrieval for only a
small proportion of packages. This results in a significant performance
gain.

This implements the performance improvement called for by #40.