bugfix: pypy3 hanging indefinitely
Pypy3 was hanging due to an attempt to add a site directory
that didn't exist. The result was a constant stat call to the
target directory.

uranium_standalone was also changed to use an exec call,
instead of a subprocess call when running uranium. This
removes the need for linking subprocess pipes, and achieves
the intended effect of becoming the uranium process