including black and pylint into uranium
vscode looks in the .uranium directory (as it is a subdirectory
of the root) for tools around linting and formatting.

to make it easier to use these tools, adding them into the
uranium package as an extra will enable it's ability
to be installed in the bootstrapper virtualenv, and in
turn be used for vscode.