moving virtualenv to .env from root
many issues arise from executing things directly in the source control
directory. one major issue has to do with the ever-growing
.gitignore file as new files get written that were previously
not so. The other disadvantage is that vscode does not
look for virtualenvs in the root, and instead looks for a subdirectory.

Moving the virtualenv to .env instead. To ensure
some level of backwards compatibility in this process,
the executables generated will be symlinked into a bin/ directory
in the previous root directory.