Fixing remaining file leaks in root directory
One goal of migrating the virtualenv directory is to prevent
leakages of files to the main directory, balooning the gitignore
and adding new entries every time an internal detail that
writes files is added.

Removing some remaining instances of using .uranium or
other files

* the history API was still writing to .uranium
* the uranium remote scripts were caching the standalone in .uranium
* executables was utilizing the root directory, which results
  in subprocesses being executed there and files such as
  coverage appearing.

As of now, the .gitignore has been reduced to six patterns.
Only four are required by uranium (2 vcver ones)