Migrating from instrumenting pip to wrapping CLI
Previously, the way that packages were installed or removed was via
a class that directly extends the internal pip classes.

This has proven to be difficult to maintain, and can result in
unexpected behavior. Pip is intended to be used a CLI, and
on top of being a large codebase, internals can be modified,
it is not intended to be a long-lived process with modified
disk stage.

The new approach is to "puppet" the CLI: enact actions using
subprocess to execute installation and removal of packages.

Updating Pip and other packaging internals, as uranium no longer
is dependent on the internal structure of the code. This was a blocker
as after pip 9.0.1 the internal structure was modified.

Removing vscode-specific extras from uranium, they now belong in uranium-plus.

As a side effect, develop eggs cannot be installed if there is a constraint
also passed into pip. To eliminate the ambiguity of a system of record of versions
in contrast to a constraint itself, splitting the constraints and versions dicts
in packages.

Removing relocatable being called. pip throws errors on develop eggs that are relocatable.
Now, virtualenv --relocatable <target_dir> should be called to have the same effect.

Due to the new strategy, it is not required to import packages such as setuptools,
as they are always installed with the .env virtualenv. Removing imports used for the
sake of ensuring availability in the app sandbox.

Two new APIs and a backwards-incompatible changes now exist on packages as a result:

* constraints: returns the constraints version object
* constraints_as_list: what the original constraints returned
* install_list: a way to group up installations of packages into a single pip command,
  improving performance.

Added code to upgrade pip on creation of a virtualenv. This ensures that, despite
using an older virtualenv version, newer pip versions will be consumed.